Added CSS styling to all role pages
They al follow the same styling conventions as the home page. I filled this in due to each folder for each role having an empty CSS styling sheet, and I just added similar styling to the homepage whilst all looking visually uniform. I understand this could have all probably been done in the one CSS sheet (the one in the source folder) but since this will be an easy change if we are required to only use one styling sheet, I can change it back quickly.